feat(plasma-tokens-utils): Update API: extract getThemeTokenDataGroups to index
