+++ utils/plasma-tokens-utils/src/generators/index.ts
@@ -4,9 +4,10 @@ export { generateTypoSystem } from './generateTypoSystem';
 export { generateThemeFromData } from './generateThemeFromData';
 export { generateFile } from './generateFile';
 export { generateThemesTokenDataGroups } from './generateThemesTokenDataGroups';
-export { getThemesTokensFallback } from './getThemesTokensFallback';
 export { generateTokenSet } from './generateTokenSet';
 export { generateThemes } from './generateThemes';
 export { generateShadowsToken } from './generateShadowsToken';
 export { generateBorderRadiusTokens } from './generateBorderRadiusTokens';
 export { generateSpacingTokens } from './generateSpacingTokens';
+export { getThemesTokensFallback } from './getThemesTokensFallback';
+export { getThemeTokenDataGroups } from './getThemeTokenDataGroups';
